Selenium/Playwright_and_pytest/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver==2.24.1
chromedriver-binary==123.0.6271.0.0
chromedriver-py==120.0.6099.109
colorama==0.4.6
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
mccabe==0.7.0
outcome==1.3.0.post0
packaging==23.2
playwright==1.41.2
pluggy==1.4.0
pycodestyle==2.11.1
pycparser==2.21
pyee==11.0.1
pyflakes==3.2.0
PySocks==1.7.1
pytest==8.0.0
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
webdriver-manager==4.0.1
wsproto==1.2.0
